Make SaveAreaView work on Android instead of custom withStatusBar wrapper
