+++ src/App.test.js
@@ -1,10 +1,9 @@
-// import React from "react";
-// import { shallow } from "enzyme";
-// import App from "./App";
+// @flow
+import React from "react";
+import { shallow } from "enzyme";
+import App from "./App";
 
-// TODO: Temporarily disabled due to https://github.com/react-navigation/react-navigation/issues/256
 it("renders correctly", () => {
-  // const output = shallow(<App />);
-  // console.log(output);
-  // expect(() => renderer.create(<App />)).not.toThrow();
+  const output = shallow(<App />);
+  expect(output).toMatchSnapshot();
 });
