+++ src/index.js
@@ -28,6 +28,7 @@ class AppWrapper extends Component<{}> {
     store.dispatch(getEvents());
     store.dispatch(loadSavedEvents());
   }
+
   render() {
     return (
       <Provider store={store}>
