+++ src/screens/EventsScreen/component.js
@@ -8,7 +8,6 @@ import type { ImageSource } from "../../data/get-asset-source";
 import EventList from "../../components/EventList";
 import FilterHeader from "../../components/ConnectedFilterHeader";
 import Loading from "../../components/Loading";
-import withStatusBar from "../../components/withStatusBar";
 import { bgColor } from "../../constants/colors";
 import {
   EVENT_DETAILS,
@@ -30,7 +29,7 @@ export type Props = {
   selectedCategories: Set<string>
 };
 
-export class EventsScreen extends PureComponent<Props> {
+class EventsScreen extends PureComponent<Props> {
   handleFilterCategoriesPress = () => {
     this.props.navigation.navigate(EVENT_CATEGORIES_FILTER);
   };
@@ -88,4 +87,4 @@ const styles = StyleSheet.create({
   }
 });
 
-export default withStatusBar(EventsScreen, {});
+export default EventsScreen;
