+++ src/screens/SavedEventListScreen/component.js
@@ -9,7 +9,6 @@ import EventList from "../../components/EventList";
 import text from "../../constants/text";
 import Loading from "../../components/Loading";
 import Header from "../../components/Header";
-import withStatusBar from "../../components/withStatusBar";
 import { bgColor } from "../../constants/colors";
 import { EVENT_DETAILS } from "../../constants/routes";
 import locale from "../../data/locale";
@@ -27,7 +26,7 @@ export type Props = {
   getAssetSource: FieldRef => ImageSource
 };
 
-export class SavedEventListScreen extends PureComponent<Props> {
+class SavedEventListScreen extends PureComponent<Props> {
   render() {
     const {
       navigation,
@@ -77,4 +76,4 @@ const styles = StyleSheet.create({
   }
 });
 
-export default withStatusBar(SavedEventListScreen, {});
+export default SavedEventListScreen;
