+++ src/screens/SupportUsScreen/index.test.js
@@ -1,7 +1,7 @@
 // @flow
 import React from "react";
 import { shallow } from "enzyme";
-import { SupportUsScreen } from ".";
+import SupportUsScreen from ".";
 
 it("renders correctly", () => {
   const output = shallow(<SupportUsScreen navigation={null} />);
